Conserta erros de logica no componente RecipeIngredients | Cria um componente Recomendations para renderizar o carrosel de recomendacoes
