+++ src/pages/Details.jsx
@@ -3,7 +3,6 @@ import React, { useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
 
 import ReactPlayer from 'react-player/youtube';
-import Carousel from 'react-multi-carousel'; // lib usada para o requisito 37
 
 import RecipeHeader from '../components/RecipeHeader';
 import RecipeInstructions from '../components/RecipeInstructions';
@@ -14,6 +13,7 @@ import {
 
 import 'react-multi-carousel/lib/styles.css';
 import '../styles/Details.css';
+import Recomendations from '../components/Recomendations';
 
 export default function Details({ match: { url, params: { id } } }) {
   const [recipe, setRecipe] = useState({});
@@ -29,36 +29,12 @@ export default function Details({ match: { url, params: { id } } }) {
     const inProgressRecipesIds = inProgressRecipesKeys
       .map((key) => Object.keys(inProgressRecipes[key])).flat();
     return inProgressRecipesIds
-      .includes(recipe.idMeal || recipe.idDrink)
+      .includes(id)
       ? 'Continuar Receita' : 'Iniciar Receita';
-  };
-
-  const showRecipeRecomendations = () => {
-    const maxRecomendations = 6;
-    const filterRecomendations = recomendations.slice(0, maxRecomendations);
-    return filterRecomendations;
   };
 
   const ingredients = Object.keys(recipe)
     .filter((ingredient) => ingredient.includes('strIngredient') && recipe[ingredient]);
-
-  const responsive = {
-    desktop: {
-      breakpoint: { max: 3000, min: 1024 },
-      items: 3,
-      slidesToSlide: 3,
-    },
-    tablet: {
-      breakpoint: { max: 1024, min: 464 },
-      items: 2,
-      slidesToSlide: 2,
-    },
-    mobile: {
-      breakpoint: { max: 464, min: 0 },
-      items: 2,
-      slidesToSlide: 2,
-    },
-  };
 
   useEffect(() => {
     async function getRecipe() {
@@ -108,32 +84,7 @@ export default function Details({ match: { url, params: { id } } }) {
         />
       )}
 
-      <Carousel
-        responsive={ responsive }
-        removeArrowOnDeviceType={ ['tablet', 'mobile'] }
-      >
-        {showRecipeRecomendations().map((recomendation, index) => (
-          <div
-            key={ recomendation.strMeal || recomendation.strDrink }
-            className={ recomendation.strMeal ? 'meal-card' : 'drink-card' }
-            data-testid={ `${index}-recomendation-card` }
-          >
-            <img
-              src={ recomendation.strMealThumb || recomendation.strDrinkThumb }
-              alt={ recomendation.strMeal || recomendation.strDrink }
-              data-testid={ `${index}-card-img` }
-            />
-            <p
-              className="category"
-            >
-              {recomendation.strAlcoholic || recomendation.strCategory}
-            </p>
-            <p data-testid={ `${index}-recomendation-title` }>
-              { recomendation.strMeal || recomendation.strDrink }
-            </p>
-          </div>
-        ))}
-      </Carousel>
+      <Recomendations recomendations={ recomendations } />
 
       {
         doneRecipes
@@ -156,6 +107,7 @@ export default function Details({ match: { url, params: { id } } }) {
 
 Details.propTypes = {
   match: shape({
+    url: string,
     params: shape({
       id: string,
     }),
